pyproject.toml
[project]
name = "abstrusegoose-wayback"
version = "0.1.0"
description = "Resumable Wayback mirror downloader for abstrusegoose.com"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "pillow>=12.2.0",
    "pytesseract>=0.3.13",
    "wayback>=0.4.5",
]
